Extend the ruleset from the npm package
We were previously using local file paths to extend the other rulesets
in this package. This works for the tests but not for someone using the other rulesets
via npm. So instead the ruleset should extend the rulesets from the npm package.

This also means we have to include the npm package itself as a `devDependency`
so that the tests will work.